+++ lib/iex-ruby-client.rb
@@ -2,6 +2,7 @@ require 'faraday'
 require 'faraday_middleware'
 require 'faraday_middleware/response_middleware'
 require 'hashie'
+require 'money_helper'
 
 require_relative 'iex/version'
 require_relative 'iex/errors'